requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.1
decorator==5.1.1
emoji==2.2.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
openai==0.27.4
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
requests==2.29.0
rfc3986==1.5.0
sniffio==1.3.0
telegram==0.0.1
tqdm==4.65.0
urllib3==1.26.15
validators==0.20.0
yarl==1.9.2
zope.interface==6.0
